Trying to implement an alt.
Just about as disgusting and terrifying as before!
I've implemented what might be a working alt, but it relies on
a thread scheduler that I've yet to build. I will probably build
a simple one just to test it out.

The alt's code is pretty horrific. Try to keep your distance.
It might get cleaned up eventually.