Added a simple thread scheduler for testing purposes
This "scheduler" just always creates new threads immediately.
There was quite a fight with Java Generics inside both the alt
and the scheduler, but I think it is ok now (there was some casting...)

The important thing is that none of the casting will be visible
to anyone using the project. I guess that is what matters... Ouch!