Added a test for the immutable list
Found one error regarding the hashcode!
Otherwise, I've added the alts to the Ops class, and I realised
that the Alt Builder holds the branches guards etc in reverse order!
I had to change the documentation a bit (I could reverse
the order, but it is a bit inefficient, and a programmer should be able
to work either way round - you'd think)